fuelLevel UI

+++ Project Vibe/Assets/Scripts/FuelBar.cs
@@ -7,12 +7,12 @@ public class FuelBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void setFuel(int fuelLevel) //moves slider UI according to fuel usage in gameplay
+    public void setFuel(float fuelLevel) //moves slider UI according to fuel usage in gameplay
     {
         slider.value = fuelLevel;
     }
 
-    public void setMaxFuel(int fuelLevel)
+    public void setMaxFuel(float fuelLevel)
     {
         slider.maxValue = fuelLevel;
         slider.value = fuelLevel;
